Fix wagtail cache header when using cache_page() (#70)
When using the `cache_page()` wrapper instead of the middleware, the
wagtail cache header is not applied correctly to pages that are served
from cache, because the wrapper function only calls the
`UpdateCacheMiddleware` if the response was not in the cache. In
particular, on requests where the response is a
`SimpleTemplateResponse`, the header incorrectly indicates a cache miss
(because in those cases the header is initially applied before the
response is stored in the cache) and in other cases the header is simply
missing (because it is initially applied after the response is stored in
the cache and therefore doesn't exist on the cached response).

This attempts to fix that, by also calling the `UpdateCacheMiddleware`
for responses that were fetched from the cache.